Clean up chat completions modules a bit
